+++ agents/action-writer-agent/actions/Sema4.ai/actions-bootstrapper/CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [1.0.4] - 2025-03-06
+
+### Changed
+
+- Dependency versions updated
+- Use `Response[T]` as return type to actions
+
+## [1.0.3] - 2025-01-09
+
+### Changed
+
+- Dependency versions updated
+
 ## [1.0.2] - 2024-10-07
 
 ### Changed