+++ agents/action-writer-agent/actions/Sema4.ai/agent-connector/CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [2.0.2] - 2025-03-06
+
+### Changed
+
+- Dependency versions updated
+
+## [2.0.1] - 2025-01-09
+
+- Update dependencies
+
 ## [2.0.0]
 
 - Support new version of Agent API