+++ agents/sales-contact-finder/actions/Sema4.ai/browsing/CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [1.2.0] - 2025-03-24
+
+### Added
+
+- External endpoint definition added to package.yaml [Read more](https://sema4.ai/docs/team-edition/marketplace/snowflake-admin#managing-external-access)
+
+## [1.1.2] - 2025-03-06
+
+### Changed
+
+- Dependency versions updated
+- Use `Response[T]` as return type to actions
+
 ## [1.1.1] - 2025-01-09
 
 ### Changed